chore: refactor folders and naming
