+++ processor-mock/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/RHEcosystemAppEng/abb-dcs-poc/processor-mock/pkg/tcp"
+	"github.com/RHEcosystemAppEng/abb-dcs-poc/processor-mock/pkg/api"
 )
 
 func main() {
-	tcp.GetMetricsOverTCP()
+	api.GetMetricsOverTCP()
 }
