+++ processor-mock/pkg/api/tcpServer.go
@@ -1,4 +1,4 @@
-package tcp
+package api
 
 import (
 	"fmt"
